refactor(server): drop dbmcp-pii and dbmcp-sql dependency edges
Move the PII row redactor (`Redactor`, `RedactionError`, `RedactionStats`)
from `dbmcp-server` to `dbmcp-pii`, where it lives next to the analyzer it
drives. `dbmcp-server` no longer depends on `dbmcp-pii` or `dbmcp-sql` —
the shared MCP `Server` wrapper, `server_info()`, opaque `Cursor`/`Pager`,
and the JSON-schema tool request/response types stand on their own.

Switch the redaction-touching tool entry points (`read_query`,
`write_query`, `explain_query`, `create_database`, `drop_database`,
`drop_table` across mysql/postgres/sqlite) to return `rmcp::model::ErrorData`
directly. A new `From<RedactionError> for ErrorData` impl in `dbmcp-pii`
lets `r.apply(&mut rows)?` propagate fail-closed redaction failures without
a manual `.map_err(...)` at twelve call sites. The `Ok(...await?)`
wrappers in each tool's `invoke` collapse to `await` (clippy
`needless_question_mark`).

`SqlError::Redaction` becomes orphan after the call-site cleanup and is
removed. `dbmcp-pii` gains direct `serde_json`, `tracing`, and `rmcp`
deps; backend crates declare their previously-transitive `dbmcp-pii` edge
explicitly. Wire format, CLI surface, and the `dbmcp::pii` tracing event
contract are unchanged. Verified by 243 unit tests, 688 Docker
integration tests, `cargo clippy --workspace --tests -- -D warnings`,
and `cargo fmt --check`.
