+++ crates/mysql/src/tools/read_query.rs
@@ -5,7 +5,6 @@ use std::borrow::Cow;
 use dbmcp_server::pagination::Pager;
 use dbmcp_server::types::{ReadQueryRequest, ReadQueryResponse};
 use dbmcp_sql::Connection as _;
-use dbmcp_sql::SqlError;
 use dbmcp_sql::StatementKind;
 use dbmcp_sql::pagination::with_limit_offset;
 use dbmcp_sql::validation::validate_read_only;
@@ -84,7 +83,7 @@ impl ToolBase for ReadQueryTool {
 
 impl AsyncTool<MysqlHandler> for ReadQueryTool {
     async fn invoke(handler: &MysqlHandler, params: Self::Parameter) -> Result<Self::Output, Self::Error> {
-        Ok(handler.read_query(params).await?)
+        handler.read_query(params).await
     }
 }
 
@@ -109,7 +108,7 @@ impl MysqlHandler {
             database,
             cursor,
         }: ReadQueryRequest,
-    ) -> Result<ReadQueryResponse, SqlError> {
+    ) -> Result<ReadQueryResponse, ErrorData> {
         let kind = validate_read_only(&query, &sqlparser::dialect::MySqlDialect {})?;
 
         let database = database.as_deref().map(str::trim).filter(|s| !s.is_empty());
