+++ crates/postgres/src/tools/drop_table.rs
@@ -74,7 +74,7 @@ impl ToolBase for DropTableTool {
 
 impl AsyncTool<PostgresHandler> for DropTableTool {
     async fn invoke(handler: &PostgresHandler, params: Self::Parameter) -> Result<Self::Output, Self::Error> {
-        Ok(handler.drop_table(params).await?)
+        handler.drop_table(params).await
     }
 }
 
@@ -97,9 +97,9 @@ impl PostgresHandler {
             table,
             cascade,
         }: DropTableRequest,
-    ) -> Result<MessageResponse, SqlError> {
+    ) -> Result<MessageResponse, ErrorData> {
         if self.config.read_only {
-            return Err(SqlError::ReadOnlyViolation);
+            return Err(SqlError::ReadOnlyViolation.into());
         }
 
         let database = database.as_deref().map(str::trim).filter(|s| !s.is_empty());
